Clean up and integrate index.
Software Boutique should be able to function as a stand alone software
management frontend when the index is not found.

The index stores brand data about the distro (or OEM). In this case,
Ubuntu MATE.

This commit makes a few structure changes to keep code stored in
controller.py, as well improve some comments.